+++ src/pcgsolver/sparsematrix.h
@@ -142,7 +142,7 @@ struct FixedSparseMatrix {
         resize(matrix.n);
         rowstart[0] = 0;
         for (unsigned int i = 0; i < n; i++) {
-            rowstart[i + 1] = rowstart[i] + matrix.index[i].size();
+            rowstart[i + 1] = rowstart[i] + (unsigned int)matrix.index[i].size();
         }
 
         value.resize(rowstart[n]);
